+++ tests/integration.rs
@@ -1,13 +1,15 @@
-use itertools::Itertools;
 use colored::Colorize;
+use itertools::Itertools;
 use sqlparser::ast;
 use qrlew::{
     Relation, With,
+    relation::display,
     sql::parse,
     io::{Database, postgresql},
 };
 #[cfg(feature = "sqlite")]
 use qrlew::io::sqlite;
+
 
 pub fn test_rewritten_eq<D: Database>(database: &mut D, query: &str) -> bool {
     let relations = database.relations();
@@ -48,28 +50,41 @@ const QUERIES: &[&str] = &[
     "SELECT 1+SUM(a), count(b) FROM table_1 WHERE d>4 GROUP BY d",
     "SELECT 1+SUM(a), count(b), d FROM table_1 GROUP BY d",
     "SELECT sum(a) FROM table_1 JOIN table_2 ON table_1.d = table_2.x",
-    "
-    WITH t1 AS (SELECT a,d FROM table_1),
+    
+    "WITH t1 AS (SELECT a,d FROM table_1),
     t2 AS (SELECT * FROM table_2)
     SELECT sum(a) FROM t1 JOIN t2 ON t1.d = t2.x",
-    "
-    WITH t1 AS (SELECT a,d FROM table_1 WHERE a>4),
+    
+    "WITH t1 AS (SELECT a,d FROM table_1 WHERE a>4),
     t2 AS (SELECT * FROM table_2)
     SELECT max(a), sum(d) FROM t1 INNER JOIN t2 ON t1.d = t2.x CROSS JOIN table_2",
-    "
-    WITH t1 AS (SELECT a,d FROM table_1),
+    
+    "WITH t1 AS (SELECT a,d FROM table_1),
     t2 AS (SELECT * FROM table_2)
     SELECT * FROM t1 INNER JOIN t2 ON t1.d = t2.x INNER JOIN table_2 ON t1.d=table_2.x ORDER BY t1.a, t2.x, t2.y, t2.z",
-    "
-    WITH t1 AS (SELECT a,d FROM table_1),
+    // Test LIMIT
+    "WITH t1 AS (SELECT a,d FROM table_1),
     t2 AS (SELECT * FROM table_2)
     SELECT * FROM t1 INNER JOIN t2 ON t1.d = t2.x INNER JOIN table_2 ON t1.d=table_2.x ORDER BY t1.a, t2.x, t2.y, t2.z LIMIT 17",
+    // Test UNION
+    "SELECT 1*a FROM table_1 UNION SELECT 1*x FROM table_2",
+    // Test no UNION with CTEs
+    "WITH t1 AS (SELECT a,d FROM table_1),
+    t2 AS (SELECT x,y FROM table_2)
+    SELECT * FROM t1",
+    // Test UNION with CTEs
+    "WITH t1 AS (SELECT 1*a, 1*d FROM table_1),
+    t2 AS (SELECT 0.1*x as a, 2*x as b FROM table_2)
+    SELECT * FROM t1 UNION SELECT * FROM t2",
+    // Some joins
+    "SELECT * FROM order_table LEFT JOIN item_table on id=order_id WHERE price>10",
     "SELECT UPPER(z) FROM table_2 LIMIT 5",
     "SELECT LOWER(z) FROM table_2 LIMIT 5",
     "SELECT CHAR_LENGTH(z) AS char_length FROM table_2 LIMIT 1",
     "SELECT POSITION('o' IN z) AS char_length FROM table_2 LIMIT 5",
 ];
 
+#[cfg(feature = "sqlite")]
 const SQLITE_QUERIES: &[&str] = &["SELECT AVG(b) as n, count(b) as d FROM table_1"];
 
 #[cfg(feature = "sqlite")]
@@ -84,8 +99,14 @@ fn test_on_sqlite() {
         assert!(test_rewritten_eq(&mut database, query));
     }
 }
+// This should work: https://www.db-fiddle.com/f/ouKSHjkEk29zWY5PN2YmjZ/10
 
-const POSTGRESQL_QUERIES: &[&str] = &["SELECT AVG(b) as n, count(b) as d FROM table_1"];
+const POSTGRESQL_QUERIES: &[&str] = &[
+    "SELECT AVG(b) as n, count(b) as d FROM table_1",
+    // Test MD5
+    "SELECT MD5(z) FROM table_2 LIMIT 10",
+    "SELECT CONCAT(x,y,z) FROM table_2 LIMIT 11",
+];
 
 #[test]
 fn test_on_postgresql() {