+++ CHANGELOG.md
@@ -6,9 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- String functions `POSITION`, `LOWER`, `UPPER` and `CHAR_LENGTH` [PR9](https://github.com/Qrlew/qrlew/pull/9)
 
+### Changed
+### Added
+
+## [0.1.2] - 2023-06-01
+### Added
+- Md5 and Concat functions
+- Protection operation
+- Set operations support
+- String functions `POSITION`, `LOWER`, `UPPER` and `CHAR_LENGTH` [PR9](https://github.com/Qrlew/qrlew/pull/9)
+- Optional handling in or [PR6](https://github.com/Qrlew/qrlew/pull/6)
+
+### Fixed
 ## [0.1.1] - 2023-05-26
 ### Changed
 - Made sqlite optional
@@ -16,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2023-05-25
 ### Added
 - First commit open source
+
 ### Fixed
 ### Changed
 ### Removed